+++ CRM/SmsChat/Upgrader.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types = 1);
+
+use CRM_SmsChat_ExtensionUtil as E;
+
+class CRM_SmsChat_Upgrader extends CRM_Extension_Upgrader_Base {
+
+}