+++ CRM/Smschat/Upgrader.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types = 1);
-
-use CRM_Smschat_ExtensionUtil as E;
-
-class CRM_Smschat_Upgrader extends CRM_Extension_Upgrader_Base {
-
-}